Extract out Ui class, add more methods to Parser class
